Refactored Enumerated Type Value Reporting
The old way of creating enumerated type values was to create an 'Enum' class
object encoding these values. We then stored in a dict in the module so that
we only ever built these classes once. Unfortunately, this doesn't work well
when using multiple different types of camera (ie, a dart and the emulator),
because they have different enumerated values for the same named enumerated
type. Instead, I've transitioned to a simplified listing of the enumerated
types by string. The enumdefs are stored in the property map - which isn't
optimal because it will get recomputed every time we want to enumerate an
enumerated value, but it is good enough for now.